move setup commands into CLI
